Forecast a model from the fable package

Produces forecasts from a trained model.

# S3 method for BINNET
forecast(object, new_data, specials = NULL, simulate = TRUE, times = 5000, ...)

Arguments

object

A model for which forecasts are required.

new_data

A tsibble containing the time points and exogenous regressors to produce forecasts for.

specials

(passed by fabletools::forecast.mdl_df()).

simulate

If TRUE, then forecast distributions are computed using simulation from a Bernoulli model.

times

The number of sample paths to use in estimating the forecast distribution when simulate = TRUE.

...

Other arguments passed to methods

Examples

melb_rain |>
  model(nn = BINNET(Wet ~ fourier(K = 1, period = "year"))) |>
  forecast(times = 10)
#> Warning: 1 error encountered for nn
#> [1] there is no package called ‘fable’
#> # A fable: 14 x 4 [1D]
#> # Key:     .model [1]
#>    .model Date          Wet .mean
#>    <chr>  <date>     <dist> <dbl>
#>  1 nn     2011-11-01     NA    NA
#>  2 nn     2011-11-02     NA    NA
#>  3 nn     2011-11-03     NA    NA
#>  4 nn     2011-11-04     NA    NA
#>  5 nn     2011-11-05     NA    NA
#>  6 nn     2011-11-06     NA    NA
#>  7 nn     2011-11-07     NA    NA
#>  8 nn     2011-11-08     NA    NA
#>  9 nn     2011-11-09     NA    NA
#> 10 nn     2011-11-10     NA    NA
#> 11 nn     2011-11-11     NA    NA
#> 12 nn     2011-11-12     NA    NA
#> 13 nn     2011-11-13     NA    NA
#> 14 nn     2011-11-14     NA    NA